// src/api/video-v1/model/timeCodes.ts
/**
 * Drtv Videos Service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TimeCodes { 
    trackUrn?: string | null;
    startTime: number;
    duration?: number | null;
    endTime: number;
    timeCodeType: string;
}